Add files via upload
This is my first project in Python.
The project comprises of the data from the client as follows.
A company has customer data that contains 8 columns of customer details and another table having name customer policy data contains the policy details of the customer. 
The company intends to offer some discount in premium for certain customers. To do that they ask their Data scientist team to get some information & perform EDA on this data set.